Renamed discovery and config packages to internal in the SqueezeBox binding. (#2633)
See openhab/static-code-analysis#100
