+++ addons/binding/org.openhab.binding.squeezebox/src/main/java/org/openhab/binding/squeezebox/handler/SqueezeBoxPlayerHandler.java
@@ -39,7 +39,7 @@ import org.eclipse.smarthome.core.types.Command;
 import org.eclipse.smarthome.core.types.RefreshType;
 import org.eclipse.smarthome.core.types.State;
 import org.openhab.binding.squeezebox.SqueezeBoxBindingConstants;
-import org.openhab.binding.squeezebox.config.SqueezeBoxPlayerConfig;
+import org.openhab.binding.squeezebox.internal.config.SqueezeBoxPlayerConfig;
 import org.openhab.binding.squeezebox.internal.utils.HttpUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
